tough: remove lifetime requirements of repository settings
Datastore: change the inner type from &'a Path to either PathBuf or
TempDir. This supports a common use case in which the user wants a
TempDir and does not want to own the lifetime of the Path.

Settings: change the datastore field to be an Option<PathBuf> instead of
&'a Path. This allows the user to specify None if they want a TempDir
to be created for them automatically, but retains the ability for the
user to specify the location of an existing directory. In the latter
case, it is often more convenient for the user to move ownership of the
Path-like object than to provide a reference.

Settings: change the two URL fields to owned Strings instead of &'a str.
This gives the user the flexibility to move temporary strings into the
Settings struct's ownership instead of creating local variables that the
user might not otherwise need.