tough: remove type parameter for transport
Removes both the lifetime and type parameters for the Repository struct.
By holding the Transport as a trait object, using Box<dyn Transport>,
the Repository struct no longer needs a type parameter.

In order to accomplish this, it was necessary to remove associated types
from the Transport trait. This leads us to return a Box<dyn Read>
and to introduce a concrete TransportError type to be used by
implementors of the Transport trait.

The upside of this change is that, for the most common known use cases
of Repository use (file and http), it is no longer necessary to make the
code generic on T: Transport. The trade-off is a slight bit of
additional work for those that would implement their own Transport
trait.

Some refactoring in the HTTP transport was called for. Though not
strictly necessary, the retry loop was so difficult to understand that
I found it necessary to factor out the inspection of the Result we get
from reqwest into its own set of functions.