tough: simplify repository load
Separates the mandatory Repository parameters from those that are
optional. The mandatory parameters remain in the Settings struct, and
the optional parameters are collected into a new Options struct which
implements the Default trait.

Changes the signature of the Repository::load function to take a
Settings struct and an Options struct. Adds a Repository::load_default
function which takes only a Settings struct and default constructs the
Options.

The net result of this breaking change is that the very common use case,
where we only want to specify the required root, metadata_url, and
targets_url parameters is supported without passing additional default
parameters.